frontend/requirements.txt
streamlit>=1.26
requests>=2.31
pandas>=2.1
numpy>=1.25
pyyaml>=6.0
altair>=5.0